backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.0
django-filter==23.3
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
